[onechat] remove agent modes
