+++ x-pack/platform/packages/shared/onechat/onechat-common/agents/index.ts
@@ -7,7 +7,6 @@
 
 export {
   AgentType,
-  AgentMode,
   oneChatDefaultAgentId,
   type AgentDescriptor,
   type AgentDefinition,
