+++ x-pack/platform/plugins/shared/onechat/server/services/agents/modes/create_handler.ts
@@ -6,7 +6,6 @@
  */
 
 import type { AgentDefinition } from '@kbn/onechat-common';
-import { AgentMode } from '@kbn/onechat-common';
 import type { AgentHandlerFn } from '@kbn/onechat-server';
 import { runAgent } from './run_agent';
 
@@ -14,17 +13,9 @@ import { runAgent } from './run_agent';
  * Create the handler function for the default onechat agent.
  */
 export const createAgentHandler = ({ agent }: { agent: AgentDefinition }): AgentHandlerFn => {
-  return async (
-    {
-      agentParams: { nextInput, conversation = [], agentMode = AgentMode.normal },
-      runId,
-      abortSignal,
-    },
-    context
-  ) => {
+  return async ({ agentParams: { nextInput, conversation = [] }, runId, abortSignal }, context) => {
     const { round } = await runAgent(
       {
-        mode: agentMode,
         nextInput,
         conversation,
         runId,
