+++ x-pack/platform/plugins/shared/onechat/server/services/agents/modes/default/convert_graph_events.ts
