+++ x-pack/platform/plugins/shared/onechat/server/services/agents/modes/default/graph.ts
