--- x-pack/platform/plugins/shared/onechat/server/services/agents/modes/researcher/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { runResearcherAgent } from './run_researcher_agent';
